update view agreement

+++ mockproject_group3/src/main/java/vn/tdt/mockproject/repository/AbstractHibernateDao.java
@@ -33,7 +33,7 @@ public abstract class AbstractHibernateDao<T extends Serializable> implements IO
 
     @Override
     public final T findOne(final long id) {
-        return (T) getCurrentSession().get(clazz, id);
+        return (T) getCurrentSession().get(clazz, (int)id);
     }
 
     @Override
